+++ hosts/legion/home.nix
@@ -5,15 +5,17 @@
   home.username = "adgai";
   home.homeDirectory = "/home/adgai";
 
-  programs.adgai.cli = {
-    gh.enable = true;
-    git.enable = true;
-    direnv.enable = true;
-    lf.enable = true;
-    tmux.enable = true;
-    shellconfig.enable = true;
+  programs.adgai = {
+    kubernetes.enable = true;
+    ui.rofi.enable = true;
+    cli = {
+      gh.enable = true;
+      git.enable = true;
+      direnv.enable = true;
+      lf.enable = true;
+      tmux.enable = true;
+      shellconfig.enable = true;
+    };
   };
-  programs.adgai.kubernetes.enable = true;
-  programs.adgai.ui.rofi.enable = true;
 
 }