+++ hosts/legion/modules/default.nix
@@ -2,15 +2,15 @@
 {
   imports = [
     ./alacritty.nix
+    # ./dconf.nix
     ./git.nix
-    ./hyperland.nix
+    # ./hyperland.nix
     ./i3.nix
     ./kubernetes.nix
     ./programs.nix
     ./qmk.nix
     ./rofi.nix
     ./wezterm.nix
-
     # ./customPkgs/python/bumblebee-status
     ./customPkgs/shell-scripts
   ];