+++ system/commons/kafka.nix
@@ -1,0 +1,6 @@
+{ pkgs, ... }:
+
+{
+  services.apache-kafka.enable = false;
+  services.zookeeper.enable = false;
+}