+++ users/common/config/nvim-config/after/plugin/mini.lua
@@ -1,3 +1,2 @@
-require("mini.starter").setup()
 require("mini.indentscope").setup()
 require("mini.cursorword").setup()