+++ users/common/config/nvim-config/lua/adgai/lsp/serverconfig.lua
@@ -4,6 +4,8 @@ local lsp = require("lsp-zero").preset({
 	manage_nvim_cmp = true,
 	suggest_lsp_servers = false,
 })
+local cmp = require("cmp")
+local cmp_action = require("lsp-zero").cmp_action()
 
 ---@diagnostic disable-next-line: unused-local
 lsp.on_attach(function(client, bufnr)
@@ -21,6 +23,8 @@ lsp.setup_servers({
 	"clangd",
 	"cssls",
 	"dockerls",
+	"pylsp",
+	-- "pyright",
 	"eslint",
 	"gopls",
 	"hls",
@@ -28,21 +32,54 @@ lsp.setup_servers({
 	"jedi_language_server",
 	"nil_ls",
 	"rnix",
-	"rust_analyzer",
 	"tailwindcss",
 	"terraform_lsp",
 	"terraformls",
 	"vimls",
 	"vuels",
-	"yamlls",
+	-- "yamlls",
 })
 
 lsp.configure("clangd", {
 	filetypes = { "c", "cpp", "objc", "objcpp", "cuda" },
 	-- cmd = { "clangd", "--background-index", "--clang-tidy", "--header-insertion=iwyu" },
 })
+require("lspconfig").yamlls.setup({
+	filetypes = { "yaml", "yaml.docker-compose", "yml" },
+	settings = {
+		yaml = {
+			schemaStore = {
+				enable = true,
+			},
+			schemas = {
+				["https://raw.githubusercontent.com/quantumblacklabs/kedro/develop/static/jsonschema/kedro-catalog-0.17.json"] = "conf/**/*catalog*",
+				["https://json.schemastore.org/github-workflow.json"] = "/.github/workflows/*",
+			},
+		},
+	},
+})
 
-lsp.setup_nvim_cmp({
+-- lsp.setup_nvim_cmp({
+-- 	sources = {
+-- 		{ name = "nvim_lsp" },
+-- 		{ name = "path" },
+-- 		{ name = "copilot" },
+-- 		{ name = "buffer", keyword_length = 3 },
+-- 		{ name = "luasnip", keyword_length = 2 },
+-- 	},
+-- })
+cmp.setup({
+	window = {
+		completion = cmp.config.window.bordered(),
+		documentation = cmp.config.window.bordered(),
+	},
+	mapping = cmp.mapping.preset.insert({
+		["<C-Space>"] = cmp.mapping.complete(),
+		["<C-f>"] = cmp_action.luasnip_jump_forward(),
+		["<C-b>"] = cmp_action.luasnip_jump_backward(),
+		["<C-u>"] = cmp.mapping.scroll_docs(-4),
+		["<C-d>"] = cmp.mapping.scroll_docs(4),
+	}),
 	sources = {
 		{ name = "nvim_lsp" },
 		{ name = "path" },
@@ -51,5 +88,18 @@ lsp.setup_nvim_cmp({
 		{ name = "luasnip", keyword_length = 2 },
 	},
 })
-lsp.nvim_workspace()
-lsp.setup()
+-- Fix lua lsp
+-- {
+--   "runtime.version": "LuaJIT",
+--   "runtime.path": [
+--     "lua/?.lua",
+--     "lua/?/init.lua"
+--   ],
+--   "diagnostics.globals": ["vim"],
+--   "workspace.checkThirdParty": false,
+--   "workspace.library": [
+--     "$VIMRUNTIME",
+--     "./lua"
+--   ]
+-- }
+--