+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 ### Feat
 
+- *****: upgrade everything for new laptop
+- **i3**: add bumblebee status
+- **wezterm**: change font to Maple Mono
+- added kafka and diffview-nvim
+- **legion**: reinstall virtualbox
+- **libvertd**: libvertd setup
+- **neovim**: add python setup
+- **programs/legion**: add doppler
+- **neovim**: remove null-ls
+- **git**: set up autoSetupRemote
+
+### Fix
+
+- **firefox**: swap firefox overlay
+- **configuration.nix**: fix firefox and nvidia driver sources
+- **flake.nix**: fix poetry overlay
+- **neovim**: fix yamlls
+- **go**: change to go 1.21
+
+## 1.2.0 (2023-09-09)
+
+### Feat
 - **legion**: added bashmount
 - moving to 23.05 for now
 - regen hardware config